@@ -1,12 +1,13 @@
 module qayyuum/go_fintech
 
-go 1.13
+go 1.22
 
 require (
-	github.com/dgrijalva/jwt-go v3.2.0+incompatible
+	github.com/golang-jwt/jwt/v5 v5.2.1
 	github.com/gorilla/mux v1.7.4
 	github.com/jinzhu/gorm v1.9.12
 	github.com/lib/pq v1.5.2
 	golang.org/x/crypto v0.0.0-20200510223506-06a226fb4e37
-	golang.org/x/tools v0.0.0-20200522201501-cb1345f3a375 // indirect
 )
+
+require github.com/jinzhu/inflection v1.0.0 // indirect
